requirements.txt
flask
tensorflow
numpy
opencv-python-headless
apache-airflow